refactor: change name from posts to blog
