+++ src/content/blog/20190513-learned-programming.md
