+++ src/content/blog/20221107-bookmarklet.md
