+++ src/content/blog/20221107-gradient-icon.md
